Keppalive issue
There is no keepalive packget if current time is very large. Caused by integer overflow.

+++ lib/include/platform_esp32_idf.h
@@ -27,5 +27,5 @@
 
 char *platform_create_id_string();
 int platform_random(int max);
-int platform_tick_get_ms();
+long long platform_tick_get_ms();
 #endif
